% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beethoven1}
\alias{beethoven1}
\title{Beethoven Example Score (Moonlight Sonata, Op. 27 No. 2)}
\format{
A data frame with note-level musical features.
}
\source{
MuseScore (public-domain Beethoven score)
}
\usage{
beethoven1
}
\description{
A sample dataset derived from a MusicXML score of Beethoven’s
Moonlight Sonata (Op. 27 No. 2), obtained from MuseScore,
included for demonstration and testing of musr functions.
}
\keyword{datasets}
