% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beethoven1_time}
\alias{beethoven1_time}
\title{Beethoven Example Time Data}
\format{
A data frame with time-based musical features.
}
\source{
MuseScore (public-domain Beethoven score)
}
\usage{
beethoven1_time
}
\description{
Time-sliced metrics derived from the Beethoven score.
}
\keyword{datasets}
