% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brahms1}
\alias{brahms1}
\title{Brahms Example Score (Geistliches Lied Op.30)}
\format{
A data frame with note-level musical features.
}
\source{
Derived from a public-domain Brahms score.
}
\usage{
brahms1
}
\description{
A sample dataset derived from a MusicXML score obtained from MuseScore,
included for demonstration and testing of musr functions.
}
\keyword{datasets}
