% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brahms1_time}
\alias{brahms1_time}
\title{Brahms Example Time Data}
\format{
A data frame with time-based musical features.
}
\source{
Derived from a public-domain Brahms score.
}
\usage{
brahms1_time
}
\description{
Time-sliced metrics derived from the Brahms score.
}
\keyword{datasets}
