% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_adds.R
\name{get_composer}
\alias{get_composer}
\title{Extract composer from a MusicXML document}
\usage{
get_composer(s)
}
\arguments{
\item{s}{An XML document object. Must be created using \code{xml2::read_xml()}.}
}
\value{
A character string containing the composer, or \code{NA} if no composer is found.
}
\description{
Retrieves the composer of a score from a MusicXML document. The function first
checks for a composer in the \code{credit} section and, if not found, falls back
to the \code{creator} element with type "composer".
}
\details{
The function prioritizes \code{//credit[credit-type='composer']/credit-words}.
If that is not present, it searches for \code{//creator[@type='composer']}.
If neither exists, \code{NA} is returned.
}
\examples{
\dontrun{
xml <- xml2::read_xml("path/to/file.xml")
get_composer(xml)
}

}
