% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_adds.R
\name{get_part_name}
\alias{get_part_name}
\title{Extract part name from a MusicXML document}
\usage{
get_part_name(s, p_id)
}
\arguments{
\item{s}{An XML document object. Must be created using \code{xml2::read_xml()}.}

\item{p_id}{A character string representing the part ID.}
}
\value{
A character string containing the part name, or \code{NA} if not found.
}
\description{
Retrieves the name of a specific part from a MusicXML document using its part ID.
}
\details{
The function searches for \code{//score-part[@id='p_id']/part-name}. If no matching
node is found, \code{NA} is returned.
}
\examples{
\dontrun{
xml <- xml2::read_xml("path/to/file.xml")
get_part_name(xml, "P1")
}

}
