% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_adds.R
\name{get_roughness}
\alias{get_roughness}
\title{Compute roughness using incon}
\usage{
get_roughness(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing roughness.
}
\description{
Calculates roughness for a numeric vector using the \code{incon} package.
}
\details{
The function validates input and calls \code{incon::roughness_hutch()}.
Requires the \code{incon} package.
}
\examples{
\dontrun{
get_roughness(c(60, 64, 67))
}

}
