% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{global_metrics}
\alias{global_metrics}
\title{Compute global roughness and harmonicity}
\usage{
global_metrics(time_df, viz = "no")
}
\arguments{
\item{time_df}{A data frame containing \code{roughness} and \code{harmonicity}.}

\item{viz}{A character string indicating whether to print a summary table ("yes" or "no").}
}
\value{
A data frame with:
\itemize{
\item \code{mean_roughness} Mean roughness
\item \code{mean_harmonicity} Mean harmonicity
}
}
\description{
Calculates mean roughness and harmonicity across all time slices.
}
\details{
If \code{viz = "yes"}, a simple table of results is printed using
\code{knitr::kable()}.
}
\examples{
\dontrun{
xml <- xml2::read_xml("path/to/file.xml")
notes_df <- parse_xml(xml)
notes_df <- expand_notes_df(notes_df, xml, "My Score")

time_df <- time_slice(notes_df)
time_df <- time_metrics(time_df)

global_metrics(time_df)
global_metrics(time_df, viz = "yes")
}

}
