% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_score}
\alias{read_score}
\title{Read a MusicXML score}
\usage{
read_score(score_path)
}
\arguments{
\item{score_path}{A character string giving the path to a MusicXML file.}
}
\value{
An XML document object containing a partwise MusicXML score.
}
\description{
Reads a MusicXML file and returns a partwise score. If the file is in
timewise format, it is automatically converted to partwise format.
}
\examples{
\dontrun{
score <- read_score("path/to/score.xml")
}

}
