% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saintsaens1}
\alias{saintsaens1}
\title{Saint-Saëns Example Score (The Carnival of the Animals: Aquarium)}
\format{
A data frame with note-level musical features.
}
\source{
Derived from a public-domain Saint-Saëns score.
}
\usage{
saintsaens1
}
\description{
A sample dataset derived from a MusicXML score obtained from MuseScore,
included for demonstration and testing of musr functions.
}
\keyword{datasets}
