% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{saintsaens1_time}
\alias{saintsaens1_time}
\title{Saint-Saëns Example Time Data}
\format{
A data frame with time-based musical features.
}
\source{
Derived from a public-domain Saint-Saëns score.
}
\usage{
saintsaens1_time
}
\description{
Time-sliced metrics derived from the Saint-Saëns score.
}
\keyword{datasets}
