% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tchai1}
\alias{tchai1}
\title{Tchaikovsky Example Score (The Nutcracker: Dance of the Sugar Plum Fairy)}
\format{
A data frame with note-level musical features.
}
\source{
Derived from a public-domain Tchaikovsky score.
}
\usage{
tchai1
}
\description{
A sample dataset derived from a MusicXML score obtained from MuseScore,
included for demonstration and testing of musr functions.
}
\keyword{datasets}
