% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tchai1_time}
\alias{tchai1_time}
\title{Tchaikovsky Example Time Data}
\format{
A data frame with time-based musical features.
}
\source{
Derived from a public-domain Tchaikovsky score.
}
\usage{
tchai1_time
}
\description{
Time-sliced metrics derived from the Tchaikovsky score.
}
\keyword{datasets}
